+++ src/helpers/input-helper.ts
@@ -18,7 +18,9 @@ export type InputCommand =
   | "cursor"
   | "secure"
   | "display_info"
-  | "list_windows";
+  | "list_windows"
+  | "screenshot"
+  | "get_ui_elements";
 
 /**
  * Execute the Swift input-helper binary with the given command and arguments.